+++ server/src/main/kotlin/fi/oph/kitu/oppijanumero/OppijanumeroService.kt
@@ -36,10 +36,3 @@ class OppijanumeroService(
             }""".trim(),
         )
 }
-
-data class YleistunnisteHaeRequest(
-    val etunimet: String,
-    val hetu: String,
-    val kutsumanimi: String,
-    val sukunimi: String,
-)
