+++ server/src/main/kotlin/fi/oph/kitu/oppijanumero/YleistunnisteHaeRequest.kt
@@ -1,0 +1,8 @@
+package fi.oph.kitu.oppijanumero
+
+data class YleistunnisteHaeRequest(
+    val etunimet: String,
+    val hetu: String,
+    val kutsumanimi: String,
+    val sukunimi: String,
+)
